cdk/functions/requirements.txt
Pillow==10.0.0
boto3==1.34.0